+++ R/regulonPruning.R
@@ -120,6 +120,7 @@ pruneRegulon <- function(regulon, expMatrix = NULL, peakMatrix = NULL, exp_assay
     }
     # choose test method
     test <- match.arg(test)
+    if(test=="chi.sq" && any(duplicated(rownames(expMatrix)))) stop("Gene names provided in 'expMatrix' are not unique.")
     message("pruning network with ", test, " tests using a regulon cutoff of ", prune_value,
         "<", regulon_cutoff)
 
@@ -523,3 +524,112 @@ chisqTest <- function(k, size, p) {
     list(p = stats::pchisq(chi, df = 1, lower.tail = FALSE), stats = chi)
 }
 
+#' Add log fold changes of gene expression to regulons
+#'
+
+#' @param expMatrix A SingleCellExperiment object or matrix containing gene expression with
+#' genes in the rows and cells in the columns
+#' @param clusters A character or integer vector of cluster or group labels for single cells
+#' @param regulon A dataframe informing the gene regulatory relationship with the ```tf``` column
+#' representing transcription factors, ```idxATAC``` corresponding to the index in the peakMatrix and
+#'  ```target``` column representing target genes
+#' @param pval.type String specifying how p-values are to be combined across pairwise comparisons for a given group/cluster.
+#' @param sig_type String specifying whether to use "FDR" or "p.value" for sig_cutoff
+#' @param logFC_condition A scalar or vector of string indicating the sample names to be compared against `logFC_ref`
+#' @param logFC_ref A scalar indicating the reference sample used to compute logFC. Default value is
+#' `rest` which is an average of all pairwise comparisons. Users can also specify a reference sample, for example, `DMSO`.
+#' @param ... additional parameters for scran::findMarkers
+#' @export
+#'
+#' @examples
+#' # create a mock singleCellExperiment object for gene expMatrixession matrix
+#' set.seed(1000)
+#' gene_sce <- scuttle::mockSCE()
+#' gene_sce <- scuttle::logNormCounts(gene_sce)
+#' rownames(gene_sce) <- paste0('Gene_',1:2000)
+#'
+#' # create a mock regulon
+#' regulon <- data.frame(tf = c(rep('Gene_1',10), rep('Gene_2',10)),
+#'                      idxATAC = sample(1:100, 20),
+#'                      target = c(paste0('Gene_', sample(3:2000,10)),
+#'                                 paste0('Gene_',sample(3:2000,10))))
+#'
+#' # filter regulon
+#' pruned.regulon <- addLogFC(expMatrix = gene_sce, clusters = gene_sce$Treatment,
+#'                                regulon = regulon,
+#'                                sig_type = "p.value")
+#'
+#' @author Xiaosai Yao
+
+addLogFC <- function(expMatrix,
+                     clusters,
+                     regulon,
+                     pval.type = c("any", "some", "all"),
+                     sig_type = c("FDR","p.value"),
+                     logFC_condition = NULL,
+                     logFC_ref = NULL,
+                     ...){
+
+  pval.type <- match.arg(pval.type)
+  sig <- match.arg(sig_type)
+
+  if (!is.null(logFC_condition)){
+    if (!all(logFC_condition %in% unique(clusters))) {
+      stop("not all conditions in logFC_conditions are found in clusters")
+    }
+  }
+
+
+  if (!is.null(logFC_ref)){
+    if (!logFC_ref %in% unique(clusters)) {
+      stop("condition in logFC_ref is not found in clusters")
+    }
+  }
+
+
+
+  if (is.null(logFC_condition)){
+    samples <- unique(clusters)
+  } else {
+    samples <- logFC_condition
+  }
+
+
+  # find differential genes
+
+  if (is.null(logFC_ref)){
+    de_list <- scran::findMarkers(x=expMatrix, groups=clusters, pval.type=pval.type, full.stats = FALSE, sorted=FALSE,...)
+
+    # combine differential genes from all clusters
+    de.df <- lapply(samples, function(sample){
+      de_genes <- as.data.frame(de_list[[sample]])
+      de_genes <- de_genes[,c(sig_type, "summary.logFC")]
+      combined_name <- paste0(sample,".vs.rest")
+      colnames(de_genes) <- c(paste0(combined_name, ".",sig_type), paste0(combined_name, ".logFC"))
+      de_genes
+    })
+
+  } else {
+    de_list <- scran::findMarkers(x=expMatrix, groups=clusters, pval.type=pval.type, full.stats = TRUE, sorted=FALSE, ...)
+
+    # combine differential genes from all clusters
+    de.df <- lapply(samples, function(sample){
+      de_genes <- as.data.frame(de_list[[sample]][,paste0("stats", ".",logFC_ref)])
+      de_genes <- de_genes[,c(paste0("log.", sig_type),"logFC")]
+      combined_name <- paste0(sample,".vs.",logFC_ref)
+      colnames(de_genes) <- c(paste0(combined_name, ".",sig_type), paste0(combined_name, ".logFC"))
+      de_genes[,1] <- 10^(de_genes[,1])
+      de_genes
+    })
+
+
+  }
+
+  de.df <- do.call(cbind, de.df)
+
+
+  # add stats
+  regulon.de <- cbind(regulon, de.df[regulon$target,])
+
+
+}